--- CoScientist/src/alembic/tools.py
@@ -1,711 +1,0 @@
-import json as _json
-import subprocess
-import textwrap
-from pathlib import Path
-
-WORKDIR = Path(
-    __import__("os").environ.get("ALEMBIC_WORKDIR", ".alembic")
-)
-MAX_BYTES = 40_000
-
-IGNORE = {
-    ".git", "__pycache__", ".eggs", "*.egg-info", "dist", "build",
-    "node_modules", ".tox", ".mypy_cache", ".pytest_cache",
-    "checkpoints", "wandb", "mlruns", ".ipynb_checkpoints",
-}
-IGNORE_EXTS = {
-    ".pyc", ".pyo", ".so", ".dylib", ".dll", ".exe",
-    ".png", ".jpg", ".jpeg", ".gif", ".svg", ".ico",
-    ".pdf", ".zip", ".tar", ".gz", ".h5", ".hdf5",
-    ".pt", ".pth", ".ckpt", ".pkl", ".npy", ".npz", ".parquet",
-}
-
-
-
-def _repo_name(repo_url: str) -> str:
-    return repo_url.rstrip("/").split("/")[-1].removesuffix(".git")
-
-
-def _repo_base(repo_url: str) -> Path:
-    """Root dir for everything related to this repo: <WORKDIR>/<repo-name>/"""
-    return WORKDIR / _repo_name(repo_url)
-
-
-def _repo_path(repo_url: str) -> Path:
-    """Where the repo is cloned: <WORKDIR>/<repo-name>/repos/"""
-    return _repo_base(repo_url) / "repos"
-
-
-def _output_dir(repo_url: str) -> Path:
-    """Where server.py, tests, .venv live: <WORKDIR>/<repo-name>/output/"""
-    return _repo_base(repo_url) / "output"
-
-
-def _reports_dir(repo_url: str) -> Path:
-    """Where .md reports live: <WORKDIR>/<repo-name>/reports/"""
-    return _repo_base(repo_url) / "reports"
-
-
-def _venv_python(out_dir: Path) -> str:
-    """Return the venv python path if it exists, else fall back to 'python'.
-
-    Uses the venv symlink path directly — do NOT resolve(), as that follows
-    the symlink to the bare uv Python binary which lacks the venv site-packages.
-    """
-    candidate = out_dir / ".venv" / "bin" / "python"
-    return str(candidate.absolute()) if candidate.exists() else "python"
-
-
-def clone_repo(repo_url: str) -> dict:
-    """Clone a GitHub repository to local disk.
-
-    Returns the local path and a flat file list for you to select from.
-
-    Example:
-        clone_repo("https://github.com/Roestlab/massformer")
-        # -> {"local_path": ".alembic/massformer/repos", "files": [...]}
-    """
-    dest = _repo_path(repo_url)
-    if not dest.exists():
-        dest.mkdir(parents=True, exist_ok=True)
-        subprocess.run(
-            ["git", "clone", "--depth=1", repo_url, str(dest)],
-            check=True, capture_output=True,
-        )
-
-    files = []
-    for p in dest.rglob("*"):
-        if p.is_file() and p.suffix not in IGNORE_EXTS:
-            rel = p.relative_to(dest)
-            if not any(part in IGNORE for part in rel.parts):
-                files.append(str(rel))
-
-    return {"local_path": str(dest), "files": sorted(files)}
-
-
-def read_file(repo_url: str, path: str) -> dict:
-    """Read a text file from the locally cloned repository.
-
-    Returns up to 40 KB of content. Do NOT use this on data files (.csv,
-    .parquet, .tsv, .json arrays) — use bash("head -n 20 <path>") instead.
-
-    Example:
-        read_file("https://github.com/Roestlab/massformer", "README.md")
-    """
-    full = _repo_path(repo_url) / path
-    if not full.exists():
-        return {"error": f"File not found: {path}."}
-    if full.suffix in IGNORE_EXTS:
-        return {"error": f"Binary/data file skipped: {path}."}
-    raw = full.read_bytes()[:MAX_BYTES]
-    return {"path": path, "content": raw.decode("utf-8", errors="replace")}
-
-
-def _glob_command(stripped: str) -> dict | None:
-    """Handle the custom ``glob <pattern>`` shortcut. Returns None if not glob."""
-    first = stripped.split()[0] if stripped else ""
-    if first != "glob":
-        return None
-    parts = stripped.split(None, 1)
-    if len(parts) < 2:
-        return {"error": "glob requires a pattern argument."}
-    pattern = parts[1]
-    matched = sorted(str(p) for p in Path("/").glob(pattern.lstrip("/")))
-    return {"matches": matched}
-
-
-def bash(command: str) -> dict:
-    """Run a shell command with a 15 s timeout.
-
-    The pipeline is intended to run inside an ephemeral container, so any
-    command line is accepted — the container is the security boundary. The
-    custom ``glob <pattern>`` shortcut is still recognised as a convenience.
-
-    Examples:
-        bash("ls .alembic/massformer/repos")
-        bash("grep -r 'def train' .alembic/massformer/repos -l")
-        bash("head -n 30 .alembic/massformer/repos/README.md")
-        bash("glob .alembic/massformer/repos/**/*.yaml")
-        bash("python -m py_compile .alembic/massformer/output/server.py && echo OK")
-    """
-    stripped = command.strip()
-    if not stripped:
-        return {"error": "empty command"}
-
-    glob_result = _glob_command(stripped)
-    if glob_result is not None:
-        return glob_result
-
-    try:
-        result = subprocess.run(
-            stripped,
-            shell=True,
-            capture_output=True,
-            text=True,
-            timeout=15,
-        )
-        output = result.stdout
-        if result.returncode != 0 and result.stderr:
-            output += "\n[stderr] " + result.stderr
-        return {"output": output[:MAX_BYTES]}
-    except subprocess.TimeoutExpired:
-        return {"error": "Command timed out after 15 seconds."}
-
-
-def bash_env(command: str) -> dict:
-    """Run a shell command with a 300 s timeout — for slow installs.
-
-    Same semantics as ``bash``, just a longer timeout so package managers
-    (pip / uv / apt-get / conda) have time to download and build.
-
-    Examples:
-        bash_env("uv venv .alembic/massformer/output/.venv --python 3.11")
-        bash_env("uv pip install --python .alembic/massformer/output/.venv/bin/python torch torchvision")
-        bash_env("pip install -r .alembic/massformer/repos/requirements.txt")
-        bash_env("which python3")
-        # System libs (container only, /var/lib/apt/lists is empty):
-        bash_env("apt-get update && apt-get install -y --no-install-recommends libpoppler-cpp-dev")
-    """
-    stripped = command.strip()
-    if not stripped:
-        return {"error": "empty command"}
-
-    glob_result = _glob_command(stripped)
-    if glob_result is not None:
-        return glob_result
-
-    try:
-        result = subprocess.run(
-            stripped,
-            shell=True,
-            capture_output=True,
-            text=True,
-            timeout=300,
-        )
-        output = result.stdout
-        if result.returncode != 0 and result.stderr:
-            output += "\n[stderr] " + result.stderr
-        return {"output": output[:MAX_BYTES]}
-    except subprocess.TimeoutExpired:
-        return {"error": "Command timed out after 300 seconds."}
-
-
-def search(repo_url: str, pattern: str) -> dict:
-    """Find files in the cloned repo matching a glob pattern.
-
-    Examples:
-        search("https://github.com/Roestlab/massformer", "**/*.yaml")
-        search("https://github.com/Roestlab/massformer", "*.sh")
-    """
-    dest = _repo_path(repo_url)
-    matched = []
-    for p in dest.glob(pattern):
-        if p.is_file() and p.suffix not in IGNORE_EXTS:
-            rel = p.relative_to(dest)
-            if not any(part in IGNORE for part in rel.parts):
-                matched.append(str(rel))
-    return {"pattern": pattern, "matches": sorted(matched)}
-
-
-def read_report(repo_url: str, report_name: str) -> dict:
-    """Read a Markdown report from this repo's reports directory.
-
-    Args:
-        repo_url:    Repository URL.
-        report_name: Filename without the .md extension: "exploration", "server",
-                     or "validation".
-
-    Example:
-        read_report("https://github.com/Roestlab/massformer", "exploration")
-        # -> {"report_path": ".alembic/massformer/reports/exploration.md", ...}
-    """
-    path = _reports_dir(repo_url) / f"{report_name}.md"
-    if not path.exists():
-        return {"error": f"No report found at {path}."}
-    return {"report_path": str(path), "content": path.read_text(encoding="utf-8")}
-
-
-def write_file(repo_url: str, relative_path: str, content: str) -> dict:
-    """Write a source file to the output directory for this repo.
-
-    Output lives at .alembic/<repo-name>/output/<relative_path>.
-
-    Examples:
-        write_file("https://github.com/Roestlab/massformer", "server.py", "...")
-        write_file("https://github.com/Roestlab/massformer", "tests/test_server.py", "...")
-        write_file("https://github.com/Roestlab/massformer", "helpers/run_analysis.py", "...")
-    """
-    dest = _output_dir(repo_url) / relative_path
-    dest.parent.mkdir(parents=True, exist_ok=True)
-    dest.write_text(content, encoding="utf-8")
-    return {"written": str(dest)}
-
-
-def read_output_file(repo_url: str, relative_path: str) -> dict:
-    """Read a file from the output directory for this repo.
-
-    Examples:
-        read_output_file("https://github.com/Roestlab/massformer", "server.py")
-        read_output_file("https://github.com/Roestlab/massformer", "tests/test_server.py")
-    """
-    full = _output_dir(repo_url) / relative_path
-    if not full.exists():
-        return {"error": f"File not found: {full}"}
-    raw = full.read_bytes()[:MAX_BYTES]
-    return {"path": str(full), "content": raw.decode("utf-8", errors="replace")}
-
-
-def update_file(repo_url: str, relative_path: str, content: str) -> dict:
-    """Overwrite a file in the output directory with corrected content.
-
-    Always write the full file — not a patch.
-
-    Examples:
-        update_file("https://github.com/Roestlab/massformer", "server.py", "...")
-        update_file("https://github.com/Roestlab/massformer", "tests/test_server.py", "...")
-    """
-    dest = _output_dir(repo_url) / relative_path
-    if not dest.exists():
-        return {"error": f"File not found: {dest}. Cannot update a file that does not exist."}
-    dest.write_text(content, encoding="utf-8")
-    return {"updated": str(dest)}
-
-
-def setup_venv(repo_url: str, packages: list[str] | None = None,
-               requirements_file: str | None = None,
-               pyproject_toml: str | None = None,
-               python_version: str | None = None) -> dict:
-    """Create a .venv in the output directory and install dependencies.
-
-    Uses `uv` when available, falls back to `python -m venv` + `pip`.
-    Always installs `mcp` and `pytest` automatically.
-
-    Args:
-        repo_url:          Repository URL.
-        packages:          Extra pip-installable package names.
-        requirements_file: Path to requirements.txt relative to cloned repo root.
-        pyproject_toml:    Path to pyproject.toml relative to cloned repo root.
-        python_version:    Python version string, e.g. "3.11".
-
-    Examples:
-        setup_venv("https://github.com/Roestlab/massformer",
-                   requirements_file="requirements.txt")
-        setup_venv("https://github.com/Roestlab/massformer",
-                   pyproject_toml="pyproject.toml", python_version="3.11")
-    """
-    out_dir  = _output_dir(repo_url)
-    out_dir.mkdir(parents=True, exist_ok=True)
-    venv_dir = out_dir / ".venv"
-    python   = str(venv_dir / "bin" / "python")
-
-    use_uv = subprocess.run(["which", "uv"], capture_output=True).returncode == 0
-
-    try:
-        if use_uv:
-            uv_venv_cmd = ["uv", "venv", str(venv_dir)]
-            if python_version:
-                uv_venv_cmd += ["--python", python_version]
-            subprocess.run(uv_venv_cmd, check=True, capture_output=True, text=True)
-        else:
-            py_bin = f"python{python_version}" if python_version else "python"
-            subprocess.run([py_bin, "-m", "venv", str(venv_dir)],
-                           check=True, capture_output=True, text=True)
-    except subprocess.CalledProcessError as e:
-        return {"success": False, "error": f"venv creation failed: {e.stderr.strip()}"}
-
-    errors = []
-
-    if requirements_file:
-        req_path = _repo_path(repo_url) / requirements_file
-        if req_path.exists():
-            try:
-                if use_uv:
-                    subprocess.run(
-                        ["uv", "pip", "install", "--python", python, "-r", str(req_path)],
-                        check=True, capture_output=True, text=True,
-                    )
-                else:
-                    subprocess.run(
-                        [str(venv_dir / "bin" / "pip"), "install", "-r", str(req_path)],
-                        check=True, capture_output=True, text=True,
-                    )
-            except subprocess.CalledProcessError as e:
-                errors.append(f"requirements install failed: {e.stderr.strip()}")
-        else:
-            errors.append(f"requirements file not found: {req_path}")
-
-    if pyproject_toml:
-        proj_path = _repo_path(repo_url) / pyproject_toml
-        if proj_path.exists():
-            try:
-                if use_uv:
-                    subprocess.run(
-                        ["uv", "pip", "install", "--python", python, "-e", str(proj_path.parent)],
-                        check=True, capture_output=True, text=True,
-                    )
-                else:
-                    subprocess.run(
-                        [str(venv_dir / "bin" / "pip"), "install", "-e", str(proj_path.parent)],
-                        check=True, capture_output=True, text=True,
-                    )
-            except subprocess.CalledProcessError as e:
-                errors.append(f"pyproject.toml install failed: {e.stderr.strip()}")
-        else:
-            errors.append(f"pyproject.toml not found: {proj_path}")
-
-    install_pkgs = ["fastmcp", "pytest", "mcp"] + (packages or [])
-    try:
-        if use_uv:
-            subprocess.run(
-                ["uv", "pip", "install", "--python", python] + install_pkgs,
-                check=True, capture_output=True, text=True,
-            )
-        else:
-            subprocess.run(
-                [str(venv_dir / "bin" / "pip"), "install"] + install_pkgs,
-                check=True, capture_output=True, text=True,
-            )
-    except subprocess.CalledProcessError as e:
-        errors.append(f"package install failed: {e.stderr.strip()}")
-
-    if errors:
-        return {"success": False, "venv": str(venv_dir), "error": "; ".join(errors)}
-    return {"success": True, "venv": str(venv_dir), "python": python}
-
-
-def check_venv_compat(repo_url: str, venv_name: str = ".venv") -> dict:
-    """Check compatibility by replaying the repo's own import statements in a venv.
-
-    Scans the cloned repo's Python files with AST, collects every unique
-    `import X` and `from X import Y` where X is an installed package, then
-    executes each statement in the venv.  This catches both ABI conflicts
-    (numpy 1 vs 2) and removed-API errors (e.g. `from transformers import AdamW`
-    removed in transformers>=4.38).
-
-    Args:
-        repo_url:  Repository URL.
-        venv_name: Directory name of the venv to check, relative to the repo
-                   output dir. Default ".venv" (the server venv). Pass
-                   ".venv-repo" to check the repo-side venv when running the
-                   two-venv layout.
-
-    Returns only failures; successful imports are omitted to keep output small.
-
-    Example:
-        check_venv_compat("https://github.com/Roestlab/massformer")
-        check_venv_compat("https://github.com/Roestlab/massformer", venv_name=".venv-repo")
-    """
-    out_dir  = _output_dir(repo_url).resolve()
-    repo_dir = _repo_path(repo_url).resolve()
-    venv_py  = out_dir / venv_name / "bin" / "python"
-    if not venv_py.exists():
-        return {"error": f"venv python not found at {venv_py}"}
-    python   = str(venv_py.absolute())
-
-    script = textwrap.dedent("""\
-        import sys, ast, importlib.metadata, json
-        from pathlib import Path
-
-        repo_path = Path(sys.argv[1])
-
-        # Build the set of top-level module names that are actually installed,
-        # so we skip repo-internal imports and uninstalled optional deps.
-        installed_roots = set()
-        for dist in importlib.metadata.distributions():
-            top = dist.read_text("top_level.txt")
-            if top:
-                for n in top.strip().splitlines():
-                    n = n.strip()
-                    if n and not n.startswith("_"):
-                        installed_roots.add(n)
-            else:
-                record = dist.read_text("RECORD") or ""
-                for line in record.splitlines():
-                    part = line.split(",")[0].strip().split("/")[0]
-                    if (not part
-                            or part.endswith((".dist-info", ".data"))
-                            or part.startswith(("_", "."))
-                            or "." in part):
-                        continue
-                    installed_roots.add(part[:-3] if part.endswith(".py") else part)
-
-        # Collect unique import statements from repo source, filtered to
-        # installed packages only (avoids false positives from repo-internal code).
-        stmts = {}  # dedup key -> statement string
-
-        for py_file in repo_path.rglob("*.py"):
-            try:
-                source = py_file.read_text(encoding="utf-8", errors="replace")
-                tree   = ast.parse(source, filename=str(py_file))
-            except Exception:
-                continue
-
-            for node in ast.walk(tree):
-                if isinstance(node, ast.Import):
-                    for alias in node.names:
-                        root = alias.name.split(".")[0]
-                        if root in installed_roots:
-                            stmt = f"import {alias.name}"
-                            stmts[stmt] = stmt
-                elif isinstance(node, ast.ImportFrom):
-                    if node.module and node.level == 0:
-                        root = node.module.split(".")[0]
-                        if root in installed_roots:
-                            names = ", ".join(a.name for a in node.names)
-                            stmt  = f"from {node.module} import {names}"
-                            key   = node.module + ":" + ",".join(
-                                sorted(a.name for a in node.names)
-                            )
-                            stmts[key] = stmt
-
-        # Execute each statement; sys.modules caches modules so heavy packages
-        # (torch, transformers) are only loaded once.
-        conflicts = {}
-        for stmt in stmts.values():
-            try:
-                exec(stmt)  # noqa: S102
-            except ImportError as e:
-                conflicts[stmt] = {"error": str(e)}
-            except Exception as e:
-                conflicts[stmt] = {"error": type(e).__name__ + ": " + str(e)}
-
-        print(json.dumps({"conflicts": conflicts, "checked": len(stmts)}))
-    """)
-
-    check_file = out_dir / "_compat_check.py"
-    check_file.write_text(script, encoding="utf-8")
-    try:
-        r = subprocess.run(
-            [python, str(check_file), str(repo_dir)],
-            capture_output=True, text=True, timeout=240,
-        )
-    finally:
-        check_file.unlink(missing_ok=True)
-
-    if r.returncode != 0:
-        return {"error": f"compat check script failed: {r.stderr.strip()[:500]}"}
-
-    try:
-        data = _json.loads(r.stdout.strip())
-    except Exception:
-        return {"error": f"could not parse compat output: {r.stdout[:300]}"}
-
-    conflicts = data.get("conflicts", {})
-    return {
-        "conflicts": conflicts,
-        "checked": data.get("checked", 0),
-        "has_conflicts": bool(conflicts),
-    }
-
-
-def validate_syntax(repo_url: str) -> dict:
-    """Check server.py for syntax errors and failed imports.
-
-    Example:
-        validate_syntax("https://github.com/Roestlab/massformer")
-    """
-    out_dir = _output_dir(repo_url).resolve()
-    server  = out_dir / "server.py"
-    python  = _venv_python(out_dir)
-    if not server.exists():
-        return {"passed": False, "stage": "syntax", "error": f"server.py not found at {server}"}
-
-    r1 = subprocess.run(
-        [python, "-m", "py_compile", str(server)],
-        capture_output=True, text=True,
-    )
-    if r1.returncode != 0:
-        return {"passed": False, "stage": "syntax", "error": r1.stderr.strip()}
-
-    load_snippet = (
-        "import importlib.util as _u, sys as _s; "
-        f"_s.path.insert(0, r'{server.parent}'); "
-        f"_spec=_u.spec_from_file_location('server', r'{server}'); "
-        "_mod=_u.module_from_spec(_spec); "
-        "_spec.loader.exec_module(_mod)"
-    )
-    r2 = subprocess.run(
-        [python, "-c", load_snippet],
-        capture_output=True, text=True, timeout=30,
-        cwd=str(server.parent),
-    )
-    if r2.returncode != 0:
-        return {"passed": False, "stage": "imports", "error": r2.stderr.strip()}
-
-    return {"passed": True}
-
-
-def run_tests(repo_url: str) -> dict:
-    """Run the pytest test suite for the generated MCP server.
-
-    Example:
-        run_tests("https://github.com/Roestlab/massformer")
-    """
-    out_dir  = _output_dir(repo_url).resolve()
-    test_dir = out_dir / "tests"
-    python   = _venv_python(out_dir)
-    if not test_dir.exists():
-        return {"passed": False, "output": f"Test directory not found: {test_dir}"}
-
-    try:
-        r = subprocess.run(
-            [python, "-m", "pytest", str(test_dir), "-v", "--tb=short", "--no-header"],
-            capture_output=True, text=True, timeout=120,
-            cwd=str(out_dir),
-        )
-    except subprocess.TimeoutExpired:
-        return {"passed": False, "output": "pytest timed out after 120 seconds."}
-
-    output = (r.stdout + r.stderr)[:MAX_BYTES]
-    return {"passed": r.returncode == 0, "output": output}
-
-
-_INVOKER_SCRIPT = textwrap.dedent('''\
-    """Invoke a single tool function from server.py with kwargs and print JSON."""
-    import importlib.util as _u
-    import json as _json
-    import os as _os
-    import sys as _sys
-    import traceback as _tb
-    from pathlib import Path as _P
-
-    _server   = _P(_os.environ["SERVER_PATH"]).resolve()
-    _toolname = _os.environ["TOOL_NAME"]
-    _args     = _json.loads(_os.environ.get("TOOL_ARGS_JSON", "{}"))
-
-    _sys.path.insert(0, str(_server.parent))
-
-    # Prevent server.py's module-level mcp.run() from blocking the import.
-    import fastmcp as _fastmcp
-    _orig_run = _fastmcp.FastMCP.run
-    _fastmcp.FastMCP.run = lambda *a, **k: None
-    try:
-        _spec = _u.spec_from_file_location("server", _server)
-        _mod  = _u.module_from_spec(_spec)
-        _spec.loader.exec_module(_mod)
-    finally:
-        _fastmcp.FastMCP.run = _orig_run
-
-    _fn = getattr(_mod, _toolname, None)
-    if _fn is None:
-        print(_json.dumps({"ok": False, "error": f"Tool {_toolname!r} not found"}))
-        raise SystemExit(0)
-    # FastMCP wraps @mcp.tool() into FunctionTool — unwrap to original callable.
-    for _attr in ("fn", "func", "__wrapped__"):
-        _inner = getattr(_fn, _attr, None)
-        if callable(_inner):
-            _fn = _inner
-            break
-
-    try:
-        _result = _fn(**_args)
-        print(_json.dumps({"ok": True, "result": _result}, default=str))
-    except Exception as _e:
-        print(_json.dumps({
-            "ok": False,
-            "error": f"{type(_e).__name__}: {_e}",
-            "traceback": _tb.format_exc(),
-        }))
-''')
-
-
-def invoke_mcp_tool(repo_url: str, tool_name: str, args: dict | None = None) -> dict:
-    """Actually invoke an @mcp.tool() function from the generated server.py.
-
-    Runs server.py inside the server venv (where fastmcp is installed),
-    looks up ``tool_name``, and calls it with ``args`` as kwargs. The
-    server's module-level ``mcp.run()`` is monkey-patched to a no-op so
-    the import does not block. FastMCP wraps tool functions in a
-    FunctionTool — we unwrap to the original callable before calling.
-
-    Use this in the validator/debugger flow to surface runtime errors
-    that pytest with mocked subprocess cannot catch (missing apt
-    packages, missing pip deps, helper-script bugs, wrong argv).
-
-    Args:
-        repo_url:  Repository URL.
-        tool_name: Exact function name (the @mcp.tool() target).
-        args:      Dict of keyword arguments to pass to the tool.
-
-    Returns:
-        On success: ``{"ok": True, "result": <whatever the tool returned>}``.
-        On error:   ``{"ok": False, "error": "<ExcName: msg>",
-                      "traceback": "<full traceback>",
-                      "stderr": "<tail of stderr>"}``.
-
-    Example:
-        invoke_mcp_tool(
-            "https://github.com/Roestlab/massformer",
-            "run_inference",
-            {"smiles_fp": "predictions/example_smiles.csv",
-             "output_fp": "predictions/demo_out.csv",
-             "custom_fp": "config/demo/demo_eval.yml",
-             "checkpoint_name": "demo",
-             "device": -1},
-        )
-    """
-    out_dir = _output_dir(repo_url).resolve()
-    server  = out_dir / "server.py"
-    venv_py = out_dir / ".venv" / "bin" / "python"
-    if not server.exists():
-        return {"ok": False, "error": f"server.py not found at {server}"}
-    if not Path(venv_py).exists():
-        return {"ok": False, "error": f"server venv python not found at {venv_py}"}
-
-    invoker = out_dir / "_invoke_tool.py"
-    invoker.write_text(_INVOKER_SCRIPT, encoding="utf-8")
-
-    env = __import__("os").environ.copy()
-    env["SERVER_PATH"]    = str(server)
-    env["TOOL_NAME"]      = tool_name
-    env["TOOL_ARGS_JSON"] = _json.dumps(args or {})
-
-    try:
-        r = subprocess.run(
-            [str(venv_py), str(invoker)],
-            capture_output=True, text=True, env=env, timeout=900,
-            cwd=str(out_dir),
-        )
-    except subprocess.TimeoutExpired:
-        invoker.unlink(missing_ok=True)
-        return {"ok": False, "error": "invocation timed out after 900 seconds"}
-    finally:
-        invoker.unlink(missing_ok=True)
-
-    # The invoker prints a single JSON line. If it crashed before printing,
-    # surface stderr so the debugger can see what failed.
-    stdout = r.stdout.strip()
-    last_line = stdout.splitlines()[-1] if stdout else ""
-    try:
-        parsed = _json.loads(last_line)
-        if not parsed.get("ok") and r.stderr:
-            parsed.setdefault("stderr", r.stderr[-2000:])
-        return parsed
-    except Exception:
-        return {
-            "ok": False,
-            "error": "could not parse invoker output",
-            "returncode": r.returncode,
-            "stdout": stdout[-1500:],
-            "stderr": r.stderr[-1500:],
-        }
-
-
-def write_report(repo_url: str, report_name: str, content: str) -> dict:
-    """Write a Markdown report to this repo's reports directory.
-
-    Args:
-        repo_url:    Repository URL.
-        report_name: Filename without the .md extension: "exploration", "server",
-                     or "validation".
-        content:     Full Markdown content to write.
-
-    Example:
-        write_report("https://github.com/Roestlab/massformer", "exploration", "# massformer...")
-        # -> {"report_path": ".alembic/massformer/reports/exploration.md"}
-    """
-    reports = _reports_dir(repo_url)
-    reports.mkdir(parents=True, exist_ok=True)
-    out = reports / f"{report_name}.md"
-    out.write_text(content, encoding="utf-8")
-    return {"report_path": str(out)}
